requirements.txt
streamlit>=1.26.4
torch==2.6.0+cu124
transformers==4.53.3
accelerate==1.9.0
numpy==1.26.4
python-docx==1.2.0
pymupdf==1.26.6